pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hoopla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.50.0",
    "nltk==3.9.1",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "pytest>=8.0.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.1",
]


[tool.setuptools]
packages = ["cli", "cli.lib"]

